hgnx2django/requirements.txt
django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
mysqlclient==2.2.0